Enable EKS clusters for QA CI pipeline

+++ qa-pipelines/tasks/cf-deploy.sh
@@ -16,19 +16,21 @@ set_uaa_sizing_params # Adds uaa sizing params to HELM_PARAMS
 # Delete legacy psp/crb, and set up new psps, crs, and necessary crbs for CAP version
 kubectl delete psp --ignore-not-found suse.cap.psp
 kubectl delete clusterrolebinding --ignore-not-found cap:clusterrole
-if semver_is_gte $(helm_chart_version) 2.15.1; then
-    kubectl delete --ignore-not-found --filename=ci/qa-tools/cap-{psp-{,non}privileged,cr-privileged-2.14.5}.yaml
-    kubectl apply --filename ci/qa-tools/cap-cr-privileged-2.15.1.yaml
-else
-    kubectl replace --force --filename=ci/qa-tools/cap-{psp-privileged,psp-nonprivileged,cr-privileged-2.14.5,crb-tests}.yaml
-fi
+if [[ ${cap_platform} != "eks" ]]; then
+    if semver_is_gte $(helm_chart_version) 2.15.1; then
+        kubectl delete --ignore-not-found --filename=ci/qa-tools/cap-{psp-{,non}privileged,cr-privileged-2.14.5}.yaml
+        kubectl apply --filename ci/qa-tools/cap-cr-privileged-2.15.1.yaml
+    else
+        kubectl replace --force --filename=ci/qa-tools/cap-{psp-privileged,psp-nonprivileged,cr-privileged-2.14.5}.yaml
+    fi
 
-kubectl replace --force --filename=ci/qa-tools/cap-crb-tests.yaml
+    kubectl replace --force --filename=ci/qa-tools/cap-crb-tests.yaml
 
-if semver_is_gte $(helm_chart_version) 2.14.5; then
-    kubectl delete --ignore-not-found --filename ci/qa-tools/cap-crb-2.13.3.yaml
-else
-    kubectl replace --filename ci/qa-tools/cap-crb-2.13.3.yaml
+    if semver_is_gte $(helm_chart_version) 2.14.5; then
+        kubectl delete --ignore-not-found --filename ci/qa-tools/cap-crb-2.13.3.yaml
+    else
+        kubectl replace --filename ci/qa-tools/cap-crb-2.13.3.yaml
+    fi
 fi
 
 echo UAA customization ...
@@ -49,7 +51,7 @@ helm install ${CAP_DIRECTORY}/helm/uaa/ \
 # Wait for UAA release
 wait_for_release uaa
 
-if [[ ${cap_platform} == "azure" ]] || [[ ${cap_platform} == "gke" ]]; then
+if [[ ${cap_platform} =~ ^azure$|^gke$|^eks$ ]]; then
     az_login
     azure_dns_clear
     azure_wait_for_lbs_in_namespace uaa
@@ -85,7 +87,7 @@ helm install ${CAP_DIRECTORY}/helm/cf/ \
 # Wait for CF release
 wait_for_release scf
 
-if [[ ${cap_platform} == "azure" ]] || [[ ${cap_platform} == "gke" ]]; then
+if [[ ${cap_platform} =~ ^azure$|^gke$|^eks$ ]]; then
     azure_wait_for_lbs_in_namespace scf
     azure_set_record_sets_for_namespace scf
 fi
